// src/api/generated/model/productoCatalogoPublico.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Ferretería API
 * API REST para gestión de ferretería con soporte multi-tenant.
      
**Características:**
- Multi-tenancy por subdominio
- Autenticación JWT
- Gestión completa de inventario
- Control de caja y ventas
- Integración SUNAT
- Auditoría completa

**Autenticación:**
Todos los endpoints (excepto /auth) requieren token JWT en header Authorization: Bearer <token>
 * OpenAPI spec version: 2.0.0
 */
import type { ProductoCatalogoPublicoDescripcion } from "./productoCatalogoPublicoDescripcion";
import type { ProductoCatalogoPublicoImagenUrl } from "./productoCatalogoPublicoImagenUrl";
import type { ProductoCatalogoPublicoMarca } from "./productoCatalogoPublicoMarca";
import type { ProductoCatalogoPublicoCategoria } from "./productoCatalogoPublicoCategoria";
import type { ProductoCatalogoPublicoSku } from "./productoCatalogoPublicoSku";

export interface ProductoCatalogoPublico {
  /** ID del producto */
  id: number;
  /** Nombre del producto */
  nombre: string;
  /** Descripción del producto */
  descripcion: ProductoCatalogoPublicoDescripcion;
  /** Precio final de venta (incluye IGV si aplica) */
  precio_venta: number;
  /** Indica si el producto tiene stock disponible */
  disponible: boolean;
  /** Stock disponible */
  stock: number;
  /** URL de la imagen principal del producto */
  imagen_url: ProductoCatalogoPublicoImagenUrl;
  /** Nombre de la marca */
  marca: ProductoCatalogoPublicoMarca;
  /** Nombre de la categoría */
  categoria: ProductoCatalogoPublicoCategoria;
  /** Código SKU del producto */
  sku: ProductoCatalogoPublicoSku;
}